+++ wetterdienst/provider/geosphere/observation/__init__.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2018-2023, earthobservations developers.
+# Distributed under the MIT License. See LICENSE for more info.
 from wetterdienst.provider.geosphere.observation.api import (
     GeosphereObservationDataset,
     GeosphereObservationParameter,